+++ hellodagger/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.airbnb.mvrx.helloDagger">
+    package="com.airbnb.mvrx.helloDagger2">
 
     <application
-        android:name=".HelloDaggerApplication"
+        android:name="com.airbnb.mvrx.helloDagger2.HelloDaggerApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.airbnb.mvrx.helloDagger2.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
