+++ hellodagger/src/main/java/com/airbnb/mvrx/helloDagger2/HelloFragment.kt
@@ -1,4 +1,4 @@
-package com.airbnb.mvrx.helloDagger
+package com.airbnb.mvrx.helloDagger2
 
 import android.os.Bundle
 import android.view.View
