+++ hellodagger/src/main/java/com/airbnb/mvrx/helloDagger2/di/AssistedViewModelFactory.kt
@@ -1,7 +1,7 @@
-package com.airbnb.mvrx.helloDagger.di
+package com.airbnb.mvrx.helloDagger2.di
 
 import com.airbnb.mvrx.MvRxState
-import com.airbnb.mvrx.helloDagger.base.BaseViewModel
+import com.airbnb.mvrx.helloDagger2.base.BaseViewModel
 
 /*
  * Serves as a supertype for AssistedInject factories in ViewModels.
